Update LoggerFactory.php
When running the same test case for different configurations like countries in the case of Panamera we need to have one report per country. This is a simple fix to name the report in wich case in the athena.json we could add a field as ENV[ATHENA_ENV_TARGET_DOMAIN] or wathever name the developer needs that might change from the execution line.