Fix to disable report on runtime.
If report params are available in athena.json file but it is important to disable it on runtime it is possible.
This fix has a caveat where the set is made on a file due to my inexperience and the fact that the settings class is instantiated multiple times in the test lifecycle. That is the reason the previous version did not work. 
I tried using also environment variables, but the ones set during runtime where lost in some point.